Assign correct end time when there's an end marker
These changes add support for setting DataPoint.TimestampRange.End
values based on another timestamp entry. This is used in cases where
one sub table is only meant to be used to indicate the end of a
DataPoint. There are cases where false, nil, or 0 represent this end
state. DataPointSetting is used to indicate what this "EndMarker' value
is meant to be.