Add tie breaker for when dissimilarity is same
This is to act as a tie breaker in situations where there are two pairs
of values that have the same dissimilarity score. In those cases, the
incorrect pair may be selected to be grouped together. We assume that
it's more likely for pairs that have more true values to be part of the
correct pair. At the same time, this value is only meant as a tie
breaker, so it's important that the change it makes to the base
dissimilarity score is very small.