Filter DataPointStrings that appear too often
It is likely that there are some DataPointStrings that appear in most
of the snapshots, but that do not actually contribute to why a snapshot
was taken at a specific moment. An example might be an aura that's
always active on a player. Even if it improves the player's stats, it
being active is not something the player is likely considering when
casting a spell. We filter these out by looking for DataPointStrings
that appear in the majority of SnapshotIds for a given category.